backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
requests==2.32.3
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlmodel

